% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestTrt.R
\name{bestTrt}
\alias{bestTrt}
\title{Best Treatment Function}
\usage{
bestTrt(mtrx)
}
\arguments{
\item{3D}{matrix (treatment x patients x simulations)}
}
\value{
2D matrix with the 3 best treatments over patients and simulations
}
\description{
function that gets a 3D matrix as an input and returns a vector with the 3 best treatments over patients and simulations
}
\examples{
foo <- array(1:125, dim = c(5,5,5))
bst <- bestTrt(foo)

}
\seealso{
\code{\link{polyasUrnFuncSimple}}

\code{\link{polyasUrnFuncAdvanced}}
}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
