% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValue.R
\name{checkValue}
\alias{checkValue}
\title{Check Value Function}
\usage{
checkValue(val)
}
\arguments{
\item{any}{numeric value}
}
\value{
updated numeric value
}
\description{
function that checks the input values if they are <1 and not NA and makes it positive
}
\examples{
foo <- -4
checkValue(foo)

}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
