% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introPrint.R
\name{introPrint}
\alias{introPrint}
\title{Intro Print Function}
\usage{
introPrint()
}
\arguments{
\item{None}{}
}
\value{
print text
}
\description{
function that prints the intro text in \code{\link{ui}}
}
\examples{

}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
