% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatio.R
\name{plotRatio}
\alias{plotRatio}
\title{Polya's Urn Plot Function}
\usage{
plotRatio(mtrx, plotNbr = 2, plotRed = 1, plotBlue = 2, plotBlack = 3)
}
\arguments{
\item{3D}{matrix (treatment x patients x simulations)}

\item{Number}{of treatments that should be plotted (default = 2)}

\item{Treatment}{index that should be represented by: Red line, Blue line, Black line (default = 1,2,3)}
}
\value{
Plot with max three treatments (y axis = ration of the treatment)  over patients (x axis) and simulations

Red line, Blue line, Black line represents each represent a treatment
}
\description{
function that gets a 3D matrix as an input and returns a plot of max 3 treatments over patients and simulations
}
\examples{
foo <- array(1:125, dim = c(5,5,5))
plotRatio(foo)

foo <- array(1:125, dim = c(5,5,5))
prinNbr <- 3
pRed <- 5
pBlue <- 4
pBlack <- 1
plotRatio(foo,prinNbr,pRed, pBlue, pBlack)

}
\seealso{
\code{\link{polyasUrnFuncSimple}}

\code{\link{polyasUrnFuncAdvanced}}
}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
