% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyasShinyRoundup.R
\name{polyasShinyRoundup}
\alias{polyasShinyRoundup}
\title{Polya's Shiny Roundup Function}
\usage{
polyasShinyRoundup()
}
\arguments{
\item{None}{}
}
\value{
opens up the shiny app and runs it
}
\description{
Function that starts the shiny app of the polya's urn simulation - the clinical trial dilemma
}
\examples{
polyasShinyRoundup()

}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
