% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyasUrnFuncAdvanced.R
\name{polyasUrnFuncAdv}
\alias{polyasUrnFuncAdv}
\title{Polya's Urn Advanced Function}
\usage{
polyasUrnFuncAdv(pat, sim, trt = 2, trtRat1 = 100, trtRat2 = 100,
  nbrBalls1 = 1, nbrBalls2 = 1, nbrRetur1 = 1, nbrRetur2 = 1,
  nbrRelap1 = 0, nbrRelap2 = 0)
}
\arguments{
\item{Number}{of patients}

\item{Number}{of simulations}

\item{Number}{of treatments (default = 2)}

\item{Success}{Rate of treatment 1 (default = 100%)}

\item{Success}{Rate of treatment 2 (default = 100%)}

\item{Number}{of Balls in the beginning for treatment 1 (default = 1)}

\item{Number}{of Balls in the beginning for treatment 2 (default = 1)}

\item{Number}{of returning balls for treatment 1 (default = 1)}

\item{Number}{of returning balls for treatment 2 (default = 1)}

\item{Number}{of relapse of treatment 1 (default = 0)}

\item{Number}{of relapse of treatment 2 (default = 0)}
}
\value{
3D matrix (treatment x patients x simulations) with the ratio of each treatment compared to all treatments for each patient and each simulation
}
\description{
function that represents the advanced polya's urn simulation setup and creates a 3D matrix (treatment x patients x simulations) to be used for the \code{\link{plotRatio}}
}
\examples{
p <- 100
s <- 2
t <- 3
plotRatio(polyasUrnFuncAvanced(p,s,t))

p <- 100
s <- 2
t <- 3
tRat1 <- 50, tRat2 <- 75
nBalls1 <- 2, nBalls2 <- 3
nRet1 <- 4, nRet2 <- 2
nRel1 <- 3
nRel2 <- 2 
plotRatio(polyasUrnFuncAdvanced(p,s,t,tRat1,tRat2,nBall1,nBalls2,nRet1,nRet2,nRel1,nRel2))

}
\seealso{
\code{\link{polyasUrnFuncSimple}}
}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
