% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyasUrnFuncSimple.R
\name{polyasUrnFuncSimple}
\alias{polyasUrnFuncSimple}
\title{Polya's Urn Simple Function}
\usage{
polyasUrnFuncSimple(pat, sim, trt = 2, trtRat = 100)
}
\arguments{
\item{pat}{Number of patients}

\item{sim}{Number of simulations}

\item{trt}{Number of treatments (default = 2)}

\item{trtRat}{Success rate of all treatments}
}
\value{
3D matrix (treatment x patients x simulations) with the ratio of each treatment compared to all treatments for each patient and each simulation
}
\description{
function that represents the simple polya's urn simulation setup and creates a 3D matrix (treatment x patients x simulations) to be used for the \code{\link{plotRatio}}
}
\examples{
p <- 100
s <- 2
t <- 3
plotRatio(polyasUrnFuncSimple(p,s,t))

}
\seealso{
\code{\link{polyasUrnFuncAdvanced}}
}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
