% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyasUrnFuncAdvanced.R
\name{polyasUrnFuncAdv}
\alias{polyasUrnFuncAdv}
\title{Polya's Urn Advanced Function}
\usage{
polyasUrnFuncAdv(pat, sim, trt = 2, trtRat1 = 100, trtRat2 = 100,
  nbrBalls1 = 1, nbrBalls2 = 1, nbrRetur1 = 1, nbrRetur2 = 1,
  nbrRelap1 = 0, nbrRelap2 = 0)
}
\arguments{
\item{pat}{Number of patients}

\item{sim}{Number of simulations}

\item{trt}{Number of treatments (default = 2)}

\item{trtRat1}{Success rate of 1. treatment}

\item{trtRat2}{Success rate of 2. treatment}

\item{nbrBalls1}{Beginn rate of balls for 1. treatment}

\item{nbrBalls2}{Beginn rate of balls for 2. treatment}

\item{nbrRetur1}{Return rate of balls for 1. treatment}

\item{nbrRetur2}{Return rate of balls for 2. treatment}

\item{nbrRelap1}{Relapse rate of balls for 1. treatment}

\item{nbrRelap2}{Relapse rate of balls for 2. treatment}
}
\value{
3D matrix (treatment x patients x simulations) with the ratio of each treatment compared to all treatments for each patient and each simulation
}
\description{
function that represents the advanced polya's urn simulation setup and creates a 3D matrix (treatment x patients x simulations) to be used for the \code{\link{plotRatio}}
Besides the same input as in \code{\link{polyasUrnFuncSimple}} additionally you have the ability to
change the treatment success rate for treatment 1 and 2, the number of starting balls (for treatment 1 and 2),
the number of return balls (for treatment 1 and 2) and the relapse rate of treatment 1 and 2.
}
\examples{
p <- 100
s <- 2
t <- 3
plotRatio(polyasUrnFuncAvanced(p,s,t))

p <- 100
s <- 2
t <- 3
tRat1 <- 50, tRat2 <- 75
nBalls1 <- 2, nBalls2 <- 3
nRet1 <- 4, nRet2 <- 2
nRel1 <- 3
nRel2 <- 2 
plotRatio(polyasUrnFuncAdvanced(p,s,t,tRat1,tRat2,nBall1,nBalls2,nRet1,nRet2,nRel1,nRel2))

}
\seealso{
\code{\link{polyasUrnFuncSimple}}
}
\author{
mitja seibold \email{mitja.seibold@student.uva.nl}
}
